% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{addition}
\alias{addition}
\title{Addition of x and y}
\usage{
addition(x = 4, y = 1)
}
\arguments{
\item{x}{is a number}

\item{y}{is another number}
}
\value{
The sum
}
\description{
Addition of x and y
}
\examples{
x + y = 5

}
